Feat: Camada Repository com DB
